personal-finance-tracker/backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Security
cryptography==42.0.0
python-decouple==3.8

# Data Processing
pandas==2.1.4
numpy==1.26.3

# File Processing
Pillow==10.2.0
python-magic==0.4.27

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# API Documentation
drf-spectacular==0.27.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
